Change DiffractionEssentialDevices to a Protocol
