+++ src/mx_bluesky/common/device_setup_plans/detector/fastcs_eiger.py
@@ -14,6 +14,7 @@ from ophyd_async.fastcs.eiger import EigerDetector
 
 from mx_bluesky.common.device_setup_plans.detector.beamline_specific import (
     BeamlineSpecificDetectorFeatures,
+    DiffractionEssentialDevices,
 )
 from mx_bluesky.common.external_interaction.callbacks.common.zocalo_callback import (
     ZocaloHWReadPayload,
@@ -21,7 +22,6 @@ from mx_bluesky.common.external_interaction.callbacks.common.zocalo_callback imp
 from mx_bluesky.common.external_interaction.callbacks.grid.grid_detect_and_scan.event_mapping import (
     HWReadDuringPayload,
 )
-from mx_bluesky.common.parameters.device_composites import DiffractionEssentialDevices
 
 
 def create_fastcs_eiger_beamline_specific(
